move batchCommit/batchReveal/batchRetrieveRewards to common, and update AVS
